requirements.txt
asgiref==3.8.1
axios==0.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coverage==7.6.11
Django==4.2.18
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
idna==3.10
importlib_metadata==8.6.1
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
psycopg2-binary==2.9.10
Pygments==2.19.1
requests==2.32.3
rich==13.9.4
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
zipp==3.21.0
